sync-npm-rev-dependencies-with-nix

+++ crates/sync_npm_rev_dependencies_with_nix/src/lib.rs
@@ -1,6 +1,7 @@
 use anyhow::Result;
 use colored::Colorize;
 use ignore::Walk;
+use npm_scaffolding_utils::guess_or_choose_package_manager;
 use parse_flake_lock::{FlakeLock, FlakeLockParseError, Node};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
@@ -27,6 +28,12 @@ pub enum SynchronizeNpmGitDependenciesWithNixError {
     #[error(transparent)]
     RegexError(#[from] regex::Error),
 
+    #[error(transparent)]
+    FileTreeError(#[from] file_tree_utils::FileTreeError),
+
+    #[error(transparent)]
+    NpmScaffoldingUtilsError(#[from] npm_scaffolding_utils::NpmScaffoldingUtilsError),
+
     #[error("Error parsing git {0} for dependency {1}")]
     ParseGitRepositoryError(String, String),
 
@@ -35,11 +42,15 @@ pub enum SynchronizeNpmGitDependenciesWithNixError {
 
     #[error("Error getting the NPM repo for dependency {0}: {1}")]
     NpmRepoError(String, String),
+
+    #[error("Error getting the NPM version for dependency {0}: {1}")]
+    NpmShowError(String, String),
 }
 
-pub fn synchronize_npm_git_dependencies_with_nix(
+pub fn synchronize_npm_rev_dependencies_with_nix(
 ) -> Result<(), SynchronizeNpmGitDependenciesWithNixError> {
-    let flake_lock = current_dir()?.join("flake.lock");
+    let current_dir = current_dir()?;
+    let flake_lock = current_dir.join("flake.lock");
 
     // Return silently if no "flake.lock" file exists
     if !flake_lock.exists() {
@@ -55,6 +66,7 @@ pub fn synchronize_npm_git_dependencies_with_nix(
         let f_name = entry.file_name().to_string_lossy();
 
         if f_name == "package.json" {
+            let mut replaced_some_dep_in_this_file = false;
             let file = File::open(entry.path())?;
             let reader = BufReader::new(file);
             let mut package_json_contents: Value = serde_json::from_reader(reader)?;
@@ -65,8 +77,8 @@ pub fn synchronize_npm_git_dependencies_with_nix(
                 for (package, dependency_source) in deps {
                     if let Value::String(dependency_source_str) = dependency_source.clone() {
                         if let Some(captures) = re.captures(&dependency_source_str) {
-                            let git_repo = get_repo(package, &dependency_source_str)?;
-                            let version = captures
+                            let git_repo = get_repo(package)?;
+                            let _version = captures
                                 .get(1)
                                 .ok_or(
                                     SynchronizeNpmGitDependenciesWithNixError::RevDependencyError(
@@ -90,24 +102,25 @@ pub fn synchronize_npm_git_dependencies_with_nix(
                                         && repo_node.locked.repo == git_repo.repo
                                         && revision != repo_node.locked.rev
                                     {
-                                        *dependency_source = Value::String(format!(
-                                            "{version}-rev.{}",
-                                            repo_node.locked.rev
-                                        ));
+                                        let new_version =
+                                            get_version(package, &repo_node.locked.rev)?;
+                                        *dependency_source =
+                                            Value::String(format!("{new_version}"));
 
                                         if !announced {
                                             announced = true;
                                             println!("");
                                             println!(
-                                        "Synchronizing npm git dependencies with the upstream nix sources..."
-                                    );
+                                                "Synchronizing npm git dependencies with the upstream nix sources..."
+                                            );
                                         }
 
                                         println!(
-                                    "  - Setting dependency \"{package}\" in file {:?} to rev \"{}\"",
-                                    entry.path(), repo_node.locked.rev
-                                );
+                                            "  - Setting dependency \"{package}\" in file {:?} to rev \"{}\"",
+                                            entry.path(), new_version
+                                        );
                                         replaced_some_dep = true;
+                                        replaced_some_dep_in_this_file = true;
                                     }
                                 }
                             }
@@ -116,16 +129,23 @@ pub fn synchronize_npm_git_dependencies_with_nix(
                 }
             }
 
-            let st = serde_json::to_string_pretty(&package_json_contents)?;
+            if replaced_some_dep_in_this_file {
+                let st = serde_json::to_string_pretty(&package_json_contents)?;
 
-            std::fs::write(entry.path(), st)?;
+                std::fs::write(entry.path(), st)?;
+            }
         }
     }
 
+    let file_tree = file_tree_utils::load_directory_into_memory(&current_dir)?;
+    let package_manager = guess_or_choose_package_manager(&file_tree)?
+        .to_string()
+        .to_lowercase();
+
     if replaced_some_dep {
-        println!("Running pnpm install...");
+        println!("Running {package_manager} install...");
         println!("");
-        Command::new("pnpm")
+        Command::new(package_manager)
             .arg("install")
             .stdout(Stdio::inherit())
             .output()?;
@@ -145,10 +165,7 @@ struct GitRepo {
     git_type: String,
 }
 
-fn get_repo(
-    npm_package: &String,
-    tag: &String,
-) -> Result<GitRepo, SynchronizeNpmGitDependenciesWithNixError> {
+fn get_repo(npm_package: &String) -> Result<GitRepo, SynchronizeNpmGitDependenciesWithNixError> {
     let output = Command::new("npm")
         .args(["repo", npm_package, "--no-browser"])
         .stderr(Stdio::null())
@@ -202,4 +219,43 @@ fn get_repo(
         repo,
         git_type,
     })
+}
+
+fn get_version(
+    npm_package: &String,
+    rev: &String,
+) -> Result<String, SynchronizeNpmGitDependenciesWithNixError> {
+    let output = Command::new("npm")
+        .args(["show", npm_package, "--json"])
+        .stderr(Stdio::null())
+        .output()?;
+    let npm_version_output = String::from_utf8(output.stdout).map_err(|err| {
+        SynchronizeNpmGitDependenciesWithNixError::NpmShowError(
+            npm_package.clone(),
+            format!("{err:?}"),
+        )
+    })?;
+
+    let mut package_json_contents: Value = serde_json::from_str(npm_version_output.as_str())?;
+
+    let Some(Value::Array(versions)) = package_json_contents.get_mut("versions") else {
+        return Err(SynchronizeNpmGitDependenciesWithNixError::NpmShowError(
+            npm_package.clone(),
+            format!("No versions returned for npm show"),
+        ));
+    };
+
+    for version in versions {
+        let Value::String(version) = version else {
+            continue;
+        };
+        if version.contains(rev) {
+            return Ok(version.clone());
+        }
+    }
+
+    Err(SynchronizeNpmGitDependenciesWithNixError::NpmShowError(
+        npm_package.clone(),
+        format!("No version found for the revision {rev}"),
+    ))
 }
